requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
colorama==0.4.3
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pipenv==2018.11.26
psycopg2==2.8.4
pylint==2.4.4
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.9
virtualenv-clone==0.5.3
whitenoise==5.0.1
wrapt==1.11.2
